Rename requestslib to librequests
