+++ socorro/lib/librequests.py
