+++ socorro/scripts/fetch_crash_data.py
@@ -8,7 +8,7 @@ import os
 import os.path
 
 from socorro.lib.datetimeutil import JsonDTEncoder
-from socorro.lib.requestslib import session_with_retries
+from socorro.lib.librequests import session_with_retries
 from socorro.scripts import FallbackToPipeAction, FlagAction, WrappedTextHelpFormatter
